# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Astronomer Management API
  version: "1.0"
  summary: Multi-cluster Kubernetes management plane
  description: |
    REST API surface for the Astronomer management plane. Covers the
    cluster onboarding wizard, agent registration, image vulnerability
    scans, kubectl shell sessions, helm catalog, and platform admin.

    Auth: all endpoints under `/api/v1/` outside `/auth/` and the
    `/internal/` namespace require either a session JWT issued by
    POST `/auth/login/` or a personal API token. Pass the token as
    `Authorization: Bearer <token>`.

    RBAC: write endpoints check the caller's permissions against the
    platform RBAC engine. Routes annotated `superuser` require
    `users.is_superuser = true`; the rest gate on resource verbs
    (clusters:update, clusters:read, etc.).
  contact:
    name: Astronomer engineering
    url: https://github.com/alphabravocompany/astronomer-go
servers:
  - url: http://astronomer.5.78.101.247.nip.io:8080
    description: Live dev (.247)
  - url: http://localhost:8080
    description: Local

# ---------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------
components:
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /api/v1/auth/login/` (session token) or
        personal API token from `/api/v1/settings/tokens/`.

  # ---------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
          required: [code, message]
      required: [error]
...
    Cluster:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, description: "RFC-1123 cluster slug" }
        display_name: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [pending, active, disconnected, error]
        environment:
          type: string
          enum: [production, staging, development, dev]
        region: { type: string }
        provider:
          type: string
          description: "eks / gke / aks / k3d / kind / other"
        distribution: { type: string, description: "vanilla / k3s / openshift / ..." }
        labels: { type: object, additionalProperties: { type: string } }
        annotations: { type: object, additionalProperties: { type: string } }
        registration_phase:
          type: string
          enum: [created, awaiting_agent, connected, provisioning, ready, failed]
        install_baseline: { type: boolean, nullable: true }
        agent_version: { type: string }
        kubernetes_version: { type: string }
        node_count: { type: integer }
        last_heartbeat: { type: string, format: date-time, nullable: true }
        is_local: { type: boolean }
        decommissioned_at: { type: string, format: date-time, nullable: true }
        agent_privilege_profile:
          type: string
          enum: [viewer, operator, admin]
        cpu_percentage: { type: number }
        memory_percentage: { type: number }
        pod_count: { type: integer }
        argocd:
          type: object
          properties:
            registered: { type: boolean }
            instance_count: { type: integer }
            cluster_secret_names:
              type: array
              items: { type: string }
            baseline_managed_by:
              type: string
              enum: [argocd, argocd_pending, helm, local, unknown]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

# Default: all routes require bearer auth.
security:
  - bearerJWT: []

# ---------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------
...
paths:
...
  /api/v1/clusters/{id}/:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Clusters]
      summary: Get cluster
      responses:
        "200":
          description: Cluster
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Cluster" }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Clusters]
      summary: Decommission cluster
      description: |
        Soft-deletes the cluster: tombstones the row + enqueues an
        async reconciler that revokes the agent's registration token,
        archives the audit log for that cluster, and prunes dependent
        rows. The cluster's pod-side resources are NOT touched — the
        operator's responsibility.
      responses:
        "204":
          description: Decommission queued
        "404":
          $ref: "#/components/responses/NotFound"
...